Ryddet i validering
