+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/bankkonto/form/index.ts
@@ -1,0 +1,5 @@
+import { UtenlandskBankkonto } from './UtenlandskBankkonto'
+import { NorskBankkonto } from './NorskBankkonto'
+import { bankkontoValidation } from '~/components/fagsystem/bankkonto/form/validation'
+
+export { UtenlandskBankkonto, NorskBankkonto, bankkontoValidation }
