+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/bankkonto/visning/NorskBankkonto.tsx
