+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/bankkonto/visning/index.ts
@@ -1,0 +1,4 @@
+import { UtenlandskBankkonto } from './UtenlandskBankkonto'
+import { NorskBankkonto } from './NorskBankkonto'
+
+export { UtenlandskBankkonto, NorskBankkonto }
