+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/tjenestepensjon/form/validation.js
@@ -1,6 +1,6 @@
 import * as Yup from 'yup'
 import { ifPresent, requiredDate, requiredString } from '~/utils/YupValidations'
-import { testDatoTom } from '~/components/fagsystem/pdlf/form/validation'
+import { testDatoTom } from '~/components/fagsystem/utils'
 
 export const validation = {
 	tp: ifPresent(
