+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/tpsmessaging/form/validation.js
@@ -1,0 +1,17 @@
+import { ifPresent, requiredString } from '~/utils/YupValidations'
+import * as Yup from 'yup'
+import { testDatoTom } from '~/components/fagsystem/utils'
+
+export const tpsMessagingValidation = {
+	tpsMessaging: ifPresent(
+		'$tpsMessaging',
+		Yup.object({
+			spraakKode: ifPresent('$tpsMessaging.spraakKode', requiredString),
+			egenAnsattDatoFom: ifPresent('$tpsMessaging.egenAnsattDatoFom', Yup.string()),
+			egenAnsattDatoTom: ifPresent(
+				'$tpsMessaging.egenAnsattDatoTom',
+				testDatoTom(Yup.string(), 'egenAnsattDatoFom')
+			),
+		})
+	),
+}
