+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/udistub/form/validation.js
@@ -1,7 +1,7 @@
 import * as Yup from 'yup'
 import _isEmpty from 'lodash/isEmpty'
 import { ifPresent, requiredBoolean, requiredString } from '~/utils/YupValidations'
-import { testDatoFom, testDatoTom } from '~/components/fagsystem/pdlf/form/validation'
+import { testDatoFom, testDatoTom } from '~/components/fagsystem/utils'
 
 const checkUndefined = (value) => {
 	if (value !== undefined) {
